+++ tests/testthat/test-coords_to_rsids.R
@@ -1,9 +1,11 @@
 test_that("coords_to_rsids works", {
   
-    dat <- dplyr::rename(echodata::BST1, SNP_old=SNP)
-    dat_annot <- echodata::coords_to_rsids(dat = dat,
-                                           SNPname = "SNP")
-    
+    dat <- echodata::BST1[seq(5),]
+    data.table::setnames(dat,"SNP","SNP_old")
+    dat_annot <- suppressWarnings(
+       coords_to_rsids(dat = dat,
+                       SNPname = "SNP")
+    )
     testthat::expect_true(methods::is(dat_annot,"data.table"))
     testthat::expect_true(!"SNP" %in% colnames(dat))
     testthat::expect_true("SNP" %in% colnames(dat_annot))